requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
prometheus_flask_exporter==0.23.1
prometheus_client==0.20.0
python-dotenv==1.0.1
